apps/features/ai/requirements.txt
# EV Charging Chatbot - Python Dependencies

# Core dependencies
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1

# Development dependencies
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0